Install policy.yaml
In OpenStack projects we are deprecating usage of policy.json and
replacing it by policy.yaml. This change ensures that the default
policy.yaml, which was added to Gnocchi recently[1], is installed.

[1] https://github.com/gnocchixyz/gnocchi/commit/0d610f6dcdfa4cb95e6832aff5054da247052a10

+++ gnocchi.spec
@@ -242,7 +242,12 @@ install -p -D -m 640 %{service}/%{service}.conf %{buildroot}%{_sysconfdir}/%{ser
 #TODO(prad): build the docs at run time, once the we get rid of postgres setup dependency
 
 # Configuration
-cp -R %{service}/rest/policy.json %{buildroot}/%{_sysconfdir}/%{service}
+if [ -f %{service}/rest/policy.json ]; then
+  cp -R %{service}/rest/policy.json %{buildroot}/%{_sysconfdir}/%{service}
+fi
+if [ -f %{service}/rest/policy.yaml ]; then
+  cp -R %{service}/rest/policy.yaml %{buildroot}/%{_sysconfdir}/%{service}
+fi
 
 # Setup directories
 install -d -m 755 %{buildroot}%{_sharedstatedir}/%{service}
